Allow min time violations in concurrent test check
The unit test that checks for concurrent termination attempts was failing on
Travis. The test assumed that the concurrent attempts would always lead to a
transaction deadlock, but it's possible (and valid) that one transaction terminates before
the other starts.

Modify the logic so that this case is not treated as an error.