Reduce load on Spinnaker
Change the way that chaosmonkey calls Spinnaker when doing terminations.

Previously, when doing terminations, chaosmonkey would query Spinnaker for the
instances in every cluster and then filter the list to the ones that were
relevant.

With this change, chaosmonkey only queries Spinnaker for instance info from
relevant clusters. This significantly reduces the number of requests that
chaosmonkey generates if the number of clusters per app is high.